requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs>=1.4.0            # via setuptools
enum34>=1.1.6             # via grpcio
grpcio-tools>=1.1.0
grpcio>=1.1.0             # via grpcio-tools
numpy>=1.12.0
packaging>=16.8           # via setuptools
protobuf>=3.2.0           # via grpcio, grpcio-tools
PyCrypto>=2.6.1
pyparsing>=2.1.10         # via packaging
six>=1.10.0               # via grpcio, packaging, protobuf, setuptools

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via protobuf
